transformer_train/requirements.txt
asgiref==3.2.7
certifi==2019.11.28
cycler==0.10.0
Django==3.0.5
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.1
prefetch-generator==1.0.1
protobuf==3.11.3
pyparsing==2.4.6
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
sqlparse==0.3.1
tensorboardX==2.0
torch-nightly==1.2.0.dev20190801+cu92
torchaudio-nightly==0.4.0.dev20190801
tqdm==4.45.0
